+++ tests/Field/Helper/FieldDefinitionHelperTest.php
@@ -4,6 +4,7 @@ namespace Tests\Torr\Storyblok\Field\Helper;
 
 use PHPUnit\Framework\TestCase;
 use Torr\Storyblok\Exception\InvalidComponentConfigurationException;
+use Torr\Storyblok\Field\Definition\AssetField;
 use Torr\Storyblok\Field\Definition\TextField;
 use Torr\Storyblok\Field\Helper\FieldDefinitionHelper;
 
@@ -42,7 +43,7 @@ final class FieldDefinitionHelperTest extends TestCase
 	/**
 	 *
 	 */
-	public function testMultiple () : void
+	public function testMultipleNames () : void
 	{
 		$this->expectException(InvalidComponentConfigurationException::class);
 		$this->expectExceptionMessage("Can't use multiple fields as admin display name");
@@ -52,6 +53,22 @@ final class FieldDefinitionHelperTest extends TestCase
 				->useAsAdminDisplayName(),
 			"two" => (new TextField("Text 2"))
 				->useAsAdminDisplayName(),
+		]);
+	}
+
+	/**
+	 *
+	 */
+	public function testMultipleImages () : void
+	{
+		$this->expectException(InvalidComponentConfigurationException::class);
+		$this->expectExceptionMessage("Can't use multiple fields as admin display image");
+
+		FieldDefinitionHelper::ensureMaximumOneAdminDisplayName([
+			"one" => (new AssetField("Image 1"))
+				->useAsAdminDisplayImage(),
+			"two" => (new AssetField("Image 2"))
+				->useAsAdminDisplayImage(),
 		]);
 	}
 }